maxot_notes/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.7.0
django-jsonfield==1.4.1
django-model-utils==4.3.1
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
hvac==1.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==22.0
Pillow==9.5.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyhcl==0.4.4
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
redis==4.5.1
requests==2.28.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
boto3==1.24.51
mysql-connector-python==8.0.32